Rework Tree objects, Update UI
Trees are now serializable objects which can persist throughout
executions. Currently, a naive check of the contents of subdirectories
confirms that a tree is valid. In future, additional checks can be made
to partially rebuild and synchronize trees. The fact that trees are now
stored removes a large portion of the startup time.

In addition, ImageNodes now lazily evaluate their children. This allows
for much faster startup, as well as lower memory footprint, since the
majority of objects in the tree are the contents of ImageNodes.
Configuration can still be achieved by specifying a path for the image
node, which is non-transient.

Event-based architecture for tree generation is removed.

The UI has been updated to display contents with a more modern and
spacious look. N-level subfolders have not been implemented yet.